refactor: update error structure
