--- src/domain/errors/index.js
@@ -1,3 +1,0 @@
-export * from "./CarError";
-export * from "./RaceError";
-export * from "./ViewError";
