+++ src/errors/domain/RaceError.js
@@ -1,4 +1,4 @@
-import { CustomError } from "./CustomError";
+import { CustomError } from "../CustomError";
 
 export class RaceError extends CustomError {
   constructor(message) {
