+++ src/errors/index.js
@@ -1,0 +1,3 @@
+export * from "./domain/CarError";
+export * from "./domain/RaceError";
+export * from "./views/ViewError";
