+++ src/errors/views/ViewError.js
@@ -1,4 +1,4 @@
-import { CustomError } from "./CustomError";
+import { CustomError } from "../CustomError";
 
 export class ViewError extends CustomError {
   constructor(message) {
